full patch

+++ PreProc/ICBC/cmip6/mod_cmip6_cnrm.F90
@@ -494,7 +494,7 @@ module mod_cmip6_cnrm
       end if
 
       tdif = idate - v%first_date
-      irec = nint((tohours(tdif)+12)/24) + 1
+      irec = int((tohours(tdif)+12)/24) + 1
 
       if ( irec > v%nrec ) then
         istatus = nf90_close(v%ncid)
